+++ src/main/java/xyz/kohara/stellarity/datagen/RecipeProvider.java
@@ -4,13 +4,16 @@ import com.mojang.serialization.Lifecycle;
 import net.fabricmc.fabric.api.datagen.v1.FabricPackOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricRecipeProvider;
 
-
 import net.minecraft.world.item.ItemStackTemplate;
-import net.minecraft.world.item.Items;
+
+import static net.minecraft.world.item.Items.*;
+
 import net.minecraft.world.item.crafting.Ingredient;
 import org.jspecify.annotations.NonNull;
 import xyz.kohara.stellarity.Stellarity;
-import xyz.kohara.stellarity.registry.StellarityItems;
+
+import static xyz.kohara.stellarity.registry.StellarityItems.*;
+
 import xyz.kohara.stellarity.registry.recipe.*;
 
 import java.util.LinkedHashMap;
@@ -57,71 +60,102 @@ public class RecipeProvider extends FabricRecipeProvider {
 		altarOfTheAccursed(output, new AltarSimpleRecipe(
 			Stellarity.id("altar_of_the_accursed/lapis_to_amethyst"),
 			new LinkedHashMap<>() {{
-				put(Ingredient.of(Items.DIAMOND), 1);
-				put(Ingredient.of(Items.LAPIS_LAZULI), 1);
+				put(Ingredient.of(DIAMOND), 1);
+				put(Ingredient.of(LAPIS_LAZULI), 1);
 			}},
-			new ItemStackTemplate(Items.AMETHYST_SHARD)
+			new ItemStackTemplate(AMETHYST_SHARD)
 		));
 
 		altarOfTheAccursed(output, new AltarSimpleRecipe(
 			Stellarity.id("altar_of_the_accursed/chorus_plating"),
 			new LinkedHashMap<>() {{
-				put(Ingredient.of(Items.IRON_INGOT), 1);
-				put(Ingredient.of(Items.POPPED_CHORUS_FRUIT), 2);
+				put(Ingredient.of(IRON_INGOT), 1);
+				put(Ingredient.of(POPPED_CHORUS_FRUIT), 2);
 			}},
-			new ItemStackTemplate(StellarityItems.CHORUS_PLATING)
+			new ItemStackTemplate(CHORUS_PLATING)
 		));
 
 		altarOfTheAccursed(output, new AltarSimpleRecipe(
 			Stellarity.id("altar_of_the_accursed/enderite_upgrade_smithing_template"),
 			new LinkedHashMap<>() {{
-				put(Ingredient.of(StellarityItems.ENDERITE_UPGRADE_SMITHING_TEMPLATE), 1);
-				put(Ingredient.of(StellarityItems.ENDERITE_SHARD), 5);
-				put(Ingredient.of(Items.PURPUR_BLOCK), 9);
+				put(Ingredient.of(ENDERITE_UPGRADE_SMITHING_TEMPLATE), 1);
+				put(Ingredient.of(ENDERITE_SHARD), 5);
+				put(Ingredient.of(PURPUR_BLOCK), 9);
 			}},
-			new ItemStackTemplate(StellarityItems.ENDERITE_UPGRADE_SMITHING_TEMPLATE, 2)
+			new ItemStackTemplate(ENDERITE_UPGRADE_SMITHING_TEMPLATE, 2)
 		));
 
 		altarOfTheAccursed(output, new AltarSimpleRecipe(
 			Stellarity.id("altar_of_the_accursed/endonomicon"),
 			new LinkedHashMap<>() {{
-				put(Ingredient.of(Items.ENCHANTED_BOOK), 1);
+				put(Ingredient.of(ENCHANTED_BOOK), 1);
 			}},
-			new ItemStackTemplate(StellarityItems.ENDONOMICON, 1)
+			new ItemStackTemplate(ENDONOMICON, 1)
 		));
 
 		altarOfTheAccursed(output, new AltarSimpleRecipe(
 			Stellarity.id("altar_of_the_accursed/satchel_of_voids"),
 			new LinkedHashMap<>() {{
-				put(Ingredient.of(Items.BUNDLE), 1);
-				put(Ingredient.of(Items.NETHER_STAR), 2);
-				put(Ingredient.of(Items.NETHERITE_INGOT), 4);
-				put(Ingredient.of(StellarityItems.ENDERITE_SHARD), 64);
-				put(Ingredient.of(StellarityItems.STARLIGHT_SOOT), 64);
+				put(Ingredient.of(BUNDLE), 1);
+				put(Ingredient.of(NETHER_STAR), 2);
+				put(Ingredient.of(NETHERITE_INGOT), 4);
+				put(Ingredient.of(ENDERITE_SHARD), 64);
+				put(Ingredient.of(STARLIGHT_SOOT), 64);
 			}},
-			new ItemStackTemplate(StellarityItems.SATCHEL_OF_VOIDS, 1)
+			new ItemStackTemplate(SATCHEL_OF_VOIDS, 1)
 		));
 
 		altarOfTheAccursed(output, new AltarSimpleRecipe(
 			Stellarity.id("altar_of_the_accursed/satchel_of_voids_alternative"),
 			new LinkedHashMap<>() {{
-				put(Ingredient.of(Items.LEATHER), 1);
-				put(Ingredient.of(Items.STRING), 1);
-				put(Ingredient.of(Items.NETHER_STAR), 2);
-				put(Ingredient.of(Items.NETHERITE_INGOT), 4);
-				put(Ingredient.of(StellarityItems.ENDERITE_SHARD), 64);
-				put(Ingredient.of(StellarityItems.STARLIGHT_SOOT), 64);
+				put(Ingredient.of(LEATHER), 1);
+				put(Ingredient.of(STRING), 1);
+				put(Ingredient.of(NETHER_STAR), 2);
+				put(Ingredient.of(NETHERITE_INGOT), 4);
+				put(Ingredient.of(ENDERITE_SHARD), 64);
+				put(Ingredient.of(STARLIGHT_SOOT), 64);
 			}},
-			new ItemStackTemplate(StellarityItems.SATCHEL_OF_VOIDS, 1)
+			new ItemStackTemplate(SATCHEL_OF_VOIDS, 1)
 		));
 
 		altarOfTheAccursed(output, new AltarUpgradeRecipe(
-			Stellarity.id("altar_of_the_accursed/test"),
-			Ingredient.of(Items.DIAMOND_HELMET),
+			Stellarity.id("altar_of_the_accursed/shulker_helmet"),
+			Ingredient.of(NETHERITE_HELMET),
 			new LinkedHashMap<>() {{
-				put(Ingredient.of(Items.NETHERITE_INGOT), 2);
+				put(Ingredient.of(ENDERITE_UPGRADE_SMITHING_TEMPLATE), 1);
+				put(Ingredient.of(SHULKER_SHELL), 4);
 			}},
-			new ItemStackTemplate(Items.NETHERITE_HELMET)
+			new ItemStackTemplate(SHULKER_HELMET)
+		));
+
+		altarOfTheAccursed(output, new AltarUpgradeRecipe(
+			Stellarity.id("altar_of_the_accursed/shulker_chestplate"),
+			Ingredient.of(NETHERITE_CHESTPLATE),
+			new LinkedHashMap<>() {{
+				put(Ingredient.of(ENDERITE_UPGRADE_SMITHING_TEMPLATE), 1);
+				put(Ingredient.of(SHULKER_SHELL), 4);
+			}},
+			new ItemStackTemplate(SHULKER_CHESTPLATE)
+		));
+
+		altarOfTheAccursed(output, new AltarUpgradeRecipe(
+			Stellarity.id("altar_of_the_accursed/shulker_leggings"),
+			Ingredient.of(NETHERITE_LEGGINGS),
+			new LinkedHashMap<>() {{
+				put(Ingredient.of(ENDERITE_UPGRADE_SMITHING_TEMPLATE), 1);
+				put(Ingredient.of(SHULKER_SHELL), 4);
+			}},
+			new ItemStackTemplate(SHULKER_LEGGINGS)
+		));
+
+		altarOfTheAccursed(output, new AltarUpgradeRecipe(
+			Stellarity.id("altar_of_the_accursed/shulker_boots"),
+			Ingredient.of(NETHERITE_BOOTS),
+			new LinkedHashMap<>() {{
+				put(Ingredient.of(ENDERITE_UPGRADE_SMITHING_TEMPLATE), 1);
+				put(Ingredient.of(SHULKER_SHELL), 4);
+			}},
+			new ItemStackTemplate(SHULKER_BOOTS)
 		));
 	}